pyproject.toml
[tool.poetry]
name = "fyp"
version = "0.1.0"
description = "Final Year Project"
authors = ["vsedov <viv.sedov@hotmail.com>"]
license = "MIT"
classifiers = ["Topic :: Utilities"]
packages = [ {include = "src"}]


[tool.poetry.dependencies]
python = ">=3.10,<3.11"
rich = "*"
requests = "*"
icecream = "*"
yapf = "^0.32.0"
numpy = "^1.23.1"
scipy = "^1.8.1"
opencv-python = "^4.6.0"
matplotlib = "^3.5.2"
tk = "^0.1.0"
pyqt5 = "^5.15.7"
torch = {version = "1.12.0", python = "3.10", platform = "linux"}
torchvision = {version = "0.13.0", python = "3.10", platform = "linux"}

tqdm = "^4.64.0"
pandas = "^1.4.3"
onnxruntime = "^1.12.1"
pytorchcv = "^0.0.67"
torchinfo = "^1.7.0"
Pillow = "^9.2.0"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"
sklearn = "^0.0"
torchsummary = "^1.5.1"
ruff = {git = "https://github.com/charliermarsh/ruff", rev = "main"}
scikit-image = "^0.19.3"
attrs = "^22.1.0"
kaggle = "^1.5.12"
pytest = "^7.2.0"
debugpy = "^1.6.3"
wandb = "^0.13.5"
plotly = "^5.11.0"
animeface = "^2.0.0"
pyflyby = "^1.8.2"
imageio = "^2.26.0"
chainer = "^7.8.1"
protobuf = "3.20"
imageio-ffmpeg = "^0.4.8"
moviepy = "^1.0.3"


[tool.poetry.dev-dependencies]
python-dotenv = "*"
isort = "^5.10.1"
coverage = "^6.3.2"
debugpy = "^1.6.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.3.23"
flake8-isort = "^4.1.1"
pep8-naming = "^0.12.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"
pytest-xdist = { version = "~=2.3.0", extras = ["psutil"] }
taskipy = "~=1.7.0"
pdoc = "^11.0.0"
pre-commit = "^2.18.1"
ipython = "^8.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python -m src"
build = "poetry build"
lint = "pre-commit run --all-files"
isort = "isort ."

# Freeze
freeze = "poetry export -f requirements.txt > requirements.txt --without-hashes"

[tool.yapf]
align_closing_bracket_with_visual_indent = "False"
allow_multiline_dictionary_keys = "True"
allow_multiline_lambdas = "False"
allow_split_before_default_or_named_assigns="False"
allow_split_before_dict_value="False"
arithmetic_precedence_indication="True"
blank_lines_around_top_level_definition="2"
# blank_lines_between_top_level_imports_and_variables="2"
blank_line_before_class_docstring="False"
blank_line_before_module_docstring="False"
blank_line_before_nested_class_or_def="True"
coalesce_brackets="True"
column_limit="120"
continuation_align_style="SPACE"
continuation_indent_width="4"
dedent_closing_brackets="False"
disable_ending_comma_heuristic="True"
each_dict_entry_on_separate_line="True"
force_multiline_dict="True"
i18n_comment="#\\..*"
i18n_function_call="N_, _"
indent_closing_brackets="False"
indent_dictionary_value="True"
indent_width="4"
join_multiple_lines="True"
no_spaces_around_selected_binary_operators="False"
spaces_around_default_or_named_assign="False"
spaces_around_dict_delimiters="False"
spaces_around_list_delimiters="False"
spaces_around_power_operator="False"
spaces_around_subscript_colon="False"
spaces_around_tuple_delimiters="False"
spaces_before_comment="2"
space_between_ending_comma_and_closing_bracket="False"
space_inside_brackets="False"
split_all_comma_separated_values="False"
split_arguments_when_comma_terminated="True"
split_before_bitwise_operator="True"
split_before_closing_bracket="True"
split_before_dict_set_generator="True"
split_before_dot="True"
split_before_expression_after_opening_paren="True"
split_before_first_argument="True"
split_before_logical_operator="True"
split_before_named_assigns="True"
split_complex_comprehension="True"
split_penalty_after_opening_bracket="300"
split_penalty_after_unary_operator="10000"
split_penalty_arithmetic_operator="300"
split_penalty_before_if_expr="0"
split_penalty_bitwise_operator="300"
split_penalty_comprehension="2100"
split_penalty_excess_character="7000"
split_penalty_for_added_line_split="30"
# split_penalty_import_names="0"
split_penalty_logical_operator="300"
use_tabs="False"

[tool.interrogate]
ignore-init-method = true
ignore-init-module = false
ignore-magic = true
ignore-semiprivate = false
ignore-private = true
ignore-property-decorators = true
ignore-module = true
fail-under = 50
exclude = ["setup.py", "docs", "build",  "_version.py", "versioneer.py"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*", "^fit$", "^transform$", "^setup"]
verbose = 1
quiet = false
whitelist-regex = []
color = true


[tool.coverage.run]
parallel = true
branch = true
source_pkgs = ["src"]
source = ["tests"]
omit = ["**/__main__.py"]

[tool.coverage.report]
show_missing = true
fail_under = 0 # temp
precision = 2
exclude_lines = ["pragma: no cover", "pass"]

[tool.isort]
multi_line_output = 6
order_by_type = false
case_sensitive = true
combine_as_imports = true
line_length = 120
atomic = true

[tool.ruff]
line-length = 150
